break out trip map into 2 components
